% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_converter.R
\name{aa_converter}
\alias{aa_converter}
\title{Arrays are awesome, but need dimnames.}
\usage{
aa_converter(arr_data, def_dn = "or", def_dnn = "orig")
}
\arguments{
\item{arr_data}{The stuff we want to work upon.}

\item{def_dn}{A character-string default value \code{"or"}, if some
dimension-names are missing, this argument will be used as a
prefix when generating them.}

\item{def_dnn}{A character-string, default value \code{"orig"}, if
the _names_ of the dimension-names are missing, this argument
will be used as a prefix when generating them.}
}
\value{
An array with a full set of dimension-names.
}
\description{
This function attempts to return an array with a full set of
dimension-names.
}
