% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arrays.R
\name{add_arrays}
\alias{add_arrays}
\title{Add arrays based on dimension-names, using recycling}
\usage{
add_arrays(.arr1, .arr2, keep_shape = FALSE)
}
\arguments{
\item{.arr1}{The first array in the sum, if \code{.arr} has the
same size (but not necessarily the same shape), the result will be
given in the shape of \code{.arr1}}

\item{.arr2}{The second array in the sum.}

\item{keep_shape}{A logig argument, default value \code{FALSE} that
only is used when recycling is needed in order to perform the sum.
If the shape of the resulting array doesn't matter, then the
default will avoid an extra permutation at the end of the
computation.}
}
\value{
When \code{.arr1} and \code{.arr2} have the same size, the
result will be the sum of them in the shape of \code{.arr1}.  If
one of them has fewer dimensions than the other, then recycling
will be performed and the result will be given in the shape of the
largest one.
}
\description{
Given two arrays with a full set of named dimension-names that
match in a proper fashion, this function will create the sum of
them based on matching dimension-names.  If necessary, an
intermediate step will be performed to properly align the arrays.
If one array has fewer dimensions than the other, then the product
will be performed by recycling of vectors.
}

