% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_index.R
\name{array_index}
\alias{array_index}
\title{Find the numerical indices in an array.}
\usage{
array_index(.bookmark, .dimnames, .check_dimnames = TRUE)
}
\arguments{
\item{.bookmark}{The caracter vector to be used for the selection.
(If a list is detected, it will be converted to a vector
first.)  This vector must have the same length as the list in
the \code{.dimnames}-argument.  If \code{.bookmark} is a named
vector, then the names must be the same as those encountered in
\code{.dimnames} (but not necessarily in the same order).}

\item{.dimnames}{The list of dimension names.  It will be assumed
that this is a named list of character vectors, and it will
also be assumed that the names (and the content of the
individual character vectors) are unique.  A test of the
validity of \code{.dimnames} will be performed when
\code{.check_dimnames} is given as \code{TRUE} (default).}

\item{.check_dimnames}{A logical value, default \code{TRUE}, that
can be used to check that \code{.dimnames} satisfies the
required uniqueness requirements.}
}
\value{
The result will be a vector of integers, which can be used
    as an alternative to \code{.bookmark} for the subsetting.
}
\description{
This function will identify the numerical indices that corresponds
to a selection based on a character vector, by comparing it to a
list representing the dimension-names to be used.
}
