% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_list_array.R
\name{array_to_list_array}
\alias{array_to_list_array}
\title{Convert a large array into a list of arrays.}
\usage{
array_to_list_array(.arr, .array_nodes = NULL, .sub_dim = NULL,
  .add_dims_attribute = FALSE)
}
\arguments{
\item{.arr}{The array to be converted into a list array.  Note that
this array must have dimension names, and that uniqueness of
the names (at the different levels) must be satisfied.  Note
that \code{.arr} itself can be a list of arrays, in which case
the function will be used on those nodes that are arrays (with
exception of place-holder nodes having a one-dimensional
array), and it will then be assumed that all the array-nodes
satisfies the requirements needed for this function to work
upon them.}

\item{.array_nodes}{Either \code{NULL} (default value) or a vector
that gives a subset of the names of the dimension-names list.
These names will be used to create a skeleton-list for the
sub-arrays to be stored, and it will be used to define the
argument \code{.sub_dim} when that is given as \code{NULL}.  If
this argument is \code{NULL} and \code{.sub_dim} is different
from \code{NULL}, then a value for \code{.array_nodes} will be
computed based on \code{.sub_dim}.  (At least one of
\code{.array_nodes} or \code{.sub_dim} must be different from
\code{NULL}.)  NB: If the value \code{character(0)} is given
(or computed), then the original array will be returned.}

\item{.sub_dim}{Either \code{NULL} (default value) or a vector that
gives a subset of the names of the dimension-names list.  These
names will be to specify the dimensions (and the order of them)
to be used for the stored sub-arrays.  It will also be used to
define the argument \code{.array_nodes} when that is given as
\code{NULL}.  If this argument is \code{NULL} and
\code{.array_nodes} is different from \code{NULL}, then a value
for \code{.sub_dim} will be computed based on \code{.sub_dim}.
(At least one of \code{.array_nodes} or \code{.sub_dim} must be
different from \code{NULL}.)}

\item{.add_dims_attribute}{Logical value, default \code{FALSE},
that can be used to return a list array with attributes that
specifies the dimension and dimension-names found at the
different nodes.}
}
\value{
An object of class \code{list_array}, constructed according
    to the specifications in the given arguments.  Note that the
    dimension names of the array-nodes all are identical, and to
    save memory those dimension-names are not stored for each
    individual array but rather stored as an attribute of the main
    result.
}
\description{
The purpose of this function is to take an array with several
dimension and adjust it into a list of arrays instead.  (It will
also work when a list of arrays are encountered).  This can e.g. be
use full if an interactive investigation requires some particular
subsets of the original array to be visualised, since an initial
conversion to list of suitable sub-arrays might speed up the
extraction of relevant slices later on.
}

