% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_names.R
\name{capture_names}
\alias{capture_names}
\title{Get the names of environments, lists and vectors.}
\usage{
capture_names(obj)
}
\arguments{
\item{obj}{The environment, list or vector under investigation.}
}
\value{
A vector containing the names from \code{obj}.  Note that
    this can have length zero, i.e. being \code{character(0)} --
    which happens when used upon an empty environment or on named,
    but empty, lists and vectors.  \code{NULL} will be returned
    when used upon a list or vector without a names attributes.
}
\description{
A helper function created to simplify the code in other functions,
primarily where an argument either can be given as an environment
or as a list/vector -- and where it's required to use the names in
some way or another.
}
