% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_spy_reports.R
\name{compare_spy_reports}
\alias{compare_spy_reports}
\title{Compare two spy-reports.}
\usage{
compare_spy_reports(first_report, second_report, ignore_these = character(0))
}
\arguments{
\item{first_report}{The first of the spy-reports.}

\item{second_report}{The second of the spy-reports.}

\item{ignore_these}{Some arguments of the targeted function might
be irrelevant for the final result, e.g. the number of cores
used in a parallel computation.  When the actual result of the
computation are the focal point of interest for our comparison,
such arguments should be included in the vector
\code{ignore_these}.}
}
\value{
Either \code{TRUE} or \code{FALSE} will be returned,
    depending on whether or not the reports where equal or not.
    Note that equality in this context focus upon the results of
    the functions, and not upon whether or not it should also
    happen that all named arguments are identical. To be precise if
    we have a function \code{f(arg)} and in addition have
    \code{name1 = name2 = 1}, then the spy-reports from \code{f(arg
    = name1)} and \code{f(arg = name2)} will be considered to be
    equal.
}
\description{
This function compares two different spy-reports (created by
\code{spy}).
}

