% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames_intersect.R
\name{dimnames_intersect}
\alias{dimnames_intersect}
\title{Intersect the dimension-names of two arrays.}
\usage{
dimnames_intersect(dimnames1, dimnames2)
}
\arguments{
\item{dimnames1}{The dimension names of the first array, the order
of the names of the dimension-names of the result will be decided
by this.}

\item{dimnames2}{The dimension names of the second array.}
}
\value{
A list with the common parts from \code{dimnames1} and
\code{dimnames2} is returned.  The result will an empty list when
no dimension-names are common, but it will also be empty if a common
dimension-name has no common values stored.
}
\description{
When a part of one array should be extracted and included into
another array, one might need to use an intersection of the
dimension-names approach to figure out the correct argument to be
used by \code{restrict_array}.  Note that the order of the
arguments does matter, since it's \code{dimnames1} that decides the
format of the resulting intersection.  Warning: The result of this
function can be an empty list, so keep in mind that your code can
deal with that if you want to use this in a function.
}
