% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames_union.R
\name{dimnames_union}
\alias{dimnames_union}
\title{Create the union of a list of dimension-names.}
\usage{
dimnames_union(dimnames_list, .restrict_list = list())
}
\arguments{
\item{dimnames_list}{The list of dimension-names, which must all
contain the same names at the top level.  The "degenerate" case
where the list has length one will be accepted, in order to
avoid specification of exceptional cases when this function is
used for programming purposes.}

\item{.restrict_list}{A list of restrictions, default value
\code{list()}, that can be used in order to shrink the
resulting union to the specified range for the given
dimensions.  Dimensions that are not included in
\code{.restrict_list} will stay untouched, and dimensions not
present in \code{dimnames_list} will be ignored.}
}
\value{
A set of dimension-names based on those from
    \code{dimnames_list}.  Any dimensions that corresponds to
    numerical values will be sorted in numerical order instead of
    lexicographical order.
}
\description{
This function will from a list of disjoint dimension-names (for a
set of sub-arrays) create a single dimension-name to be used for an
array that can be used to store the stuff from the sub-arrays.
}

