% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\title{Create error-messages.}
\usage{
error(..., .argument = NULL, n = 1, fun_stack = TRUE, quoted = FALSE,
  width = 80, indent = 4, exdent = 4, call. = FALSE, domain = NULL)
}
\arguments{
\item{...}{A collection of vectors or lists containing the stuff
that first will be inserted into \code{code_str} and then given
to \code{stop}.  Any lists will (after a sanity-check) be split
into separate vectors; and each vector will be transformed to
its separate paragraph, which will be separated by a blank
line.  Note that any formatting (like special characters for
line breaks or tabulators) will be ignored.}

\item{.argument}{The name(s) of the argument(s) that triggered the
error.  The default value for this argument is \code{NULL},
which suppress this part of the message.}

\item{n}{A non-negative integer that select which function to
report the error for.  The default value \code{1} selects the
function from which \code{error} was called, whereas higher
values goes further up the call-stack (useful when
sanity-checking of arguments is taken care of by a separate
helper-function).  Use \code{0} to suppress information about
the functions. generations to report.}

\item{fun_stack}{Logical value, default \code{TRUE}, that decides
if a presentation of the full call-stack from the top-level to
the function specified by \code{n}.  Note that this argument
will be reset to \code{FALSE} if \code{n} is \code{0}.}

\item{quoted}{Logical value, default \code{FALSE}. This can be used
to return a quoted version of the error.}

\item{width}{The value to be given to the \code{width} argument of
\code{str_wrap}.  Default value \code{80}.}

\item{indent}{The value to be given to the \code{indent} argument
of \code{str_wrap}.  Default value \code{4}.}

\item{exdent}{The value to be given to the \code{exdent} argument
of \code{str_wrap}.  Default value \code{4}.}

\item{call.}{The value to be given to the \code{call.}-argument of
\code{stop}: logical, indicating if the call should become part
of the error message.  The default value for this function is
\code{FALSE}.}

\item{domain}{The value to be given to the \code{domain}-argument
of \code{stop}: see \code{gettext}.  If \code{NA}, messages
will not be translated.  The default value of this argument is
\code{NULL}.}
}
\value{
The effect of this function will be that the information
    given to the \code{dotsMethods}-argument will be presented by
    means of \code{str_wrap} and \code{stop}, with variations
    depending on the values given to the other arguments.  Note
    that nothing will be done if no arguments are given to
    \code{...}.  If the function is used in the global workspace
    (happens under the development phase when writing a new
    function), then an adjusted message will be given with regard
    to the effect of the arguments \code{fun_stack} and \code{n}.
    The argument \code{quoted} can be used to return a quoted
    expression of the error, which might be useful if one would
    like to figure out if several errors was triggered by faulty
    arguments.
}
\description{
This function wraps around \code{stop} and \code{str_wrap} from the
\code{stringr}-package, and is a minor helper with regard to
including detailed error-messages in the code without to much fuzz.
It's also possible to get a quoted result.
}

