% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_created.R
\name{file_created}
\alias{file_created}
\title{Help function: File created}
\usage{
file_created(file_name)
}
\arguments{
\item{file_name}{A character string giving the name of the file
that was created.}
}
\description{
In order to see progress, I fancy seeing messages about files that
has been created.  This is not intended to be exported, but while
developing this stuff it is nice to have it available.
}
