% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_info.R
\name{fun_info}
\alias{fun_info}
\title{Find package and formals for exported and non-exported functions.}
\usage{
fun_info(target)
}
\arguments{
\item{target}{This can either be the function itself, or it can be
the name of it (as a symbol or a character-string), or it can e
a call or an expression that contains the desired function.}
}
\value{
A list containing the name of the function (as a
    character-string), the package it belongs to (or NULL) if no
    package was detected, and the arguments of the function.  If
    the function was given as a call or expression, any values
    specified there will be included, otherwise the default values
    will be used instead.
}
\description{
This function takes care of the extraction of information related
to a function, i.e. the name of the function, the package(s) it
belongs to (if it does belong to any packages), whether or not it
is exported from these packages, and the formals of these
functions.  If the function also can be found in \code{.GlobalEnv},
then it will be registered with package given as \code{"NULL"}.
}

