% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_status.R
\name{fun_status}
\alias{fun_status}
\title{Function status, exported or internal.}
\usage{
fun_status(.fun, .package)
}
\arguments{
\item{.fun}{The name of the function to investigate, given as a
character-string.}

\item{.package}{The name of the package \code{.fun} originates
from.}
}
\value{
Eiter "exported" or "internal", to reveal the status of the
    package.
}
\description{
A little helper to investigate if a given function from a given
package belongs to the part that has been exported or imported.
Note that this function assumes that the namespace of the package
has been loaded, and that the function is known to exist inside of
it.  The reason for this simplifying assumptions is due to the way
this function is used at the moment of writing.
}
