% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_testing.R
\name{is_testing}
\alias{is_testing}
\title{Test an object for several properties.}
\usage{
is_testing(obj, .is = NULL, .is_criteria = c("any", "all"),
  .is_not = NULL, .length = NULL, .env = parent.frame(),
  .loop = FALSE)
}
\arguments{
\item{obj}{The object to be tested}

\item{.is}{A vector with the specification of what to test.  Note
that this should be the part after the \code{.} in the tests
you would like to try, e.g.\ \code{c("character", "vector"))}
will trigger the tests \code{is.character} and
\code{is.vector}.  The default value \code{NULL} will trigger
that no tests are performed.}

\item{.is_criteria}{One of the values \code{c("any", "all")} that
will decide whether or not \code{any} or \code{all} will be used
when testing the criteria given in \code{.is}.  The default for
this argument is \code{"any"}.}

\item{.is_not}{Similar to \code{.is}, but used to specify
properties we don't want \code{obj} to have.  The default value
\code{NULL} will trigger that no such negative tests are
performed.  Note that all the negative tests must be passed in
order for this function to return \code{TRUE}.}

\item{.length}{An integer that can be used to specify the length
\code{obj} should have.  The default value \code{NULL} will
allow \code{obj} to have any length.  Use \code{"!0"} if it's a
requirement that \code{obj} should have positive length.}

\item{.env}{The environment in which the testing should be done.
The default value is \code{parent.frame()}, but other arguments
might be needed during the investigation.}

\item{.loop}{A logical argument, default \code{FALSE} that can be
used if it's of interest to investigate all the elements of a
given vector/list.  Note that it's quite likely that if symbols
are included in a list/vector, then it's hard to use any of the
apply-functions to loop over it.}
}
\value{
The result will be \code{TRUE} or \code{FALSE} depending on
    whether or not the specified criteria turned out to be
    satisfied or not.  Diagnostic attributes will be added to the
    result in order to produce error-messages if that should be of
    interest.
}
\description{
This function can be used if you would like to test if an object
satisfies a bunch of criteria.
}
\details{
WARNING: This function, in it's present incarnation, might
    need some of its arguments quoted if it's going to be used
    together with e.g. `vapply`.  It might thus be preferable to
    adjust the setup so it can investigate vectors and lists
    itself.  I think that task should be feasible, but it might
    require an additional argument to do this (that is to say, an
    argument that can be used when the function calls itself
    iteratively).
}
