% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_array_dims.R
\name{list_array_dims}
\alias{list_array_dims}
\title{Join similar lists along array-nodes.}
\usage{
list_array_dims(.list)
}
\arguments{
\item{.list}{A list (presumably with some nodes containing arrays)
where it is of interest to extract the dimension and dimension
names of these arrays.}
}
\value{
The list given to \code{.list} will be returned with
    additional attributes that specifies the specifyeis A single
    list where the dimension names of the array-nodes replaces the
    arrays, and where all other stuff has been dropped.
}
\description{
This function simplifies the task of working with list where some
of the nodes are arrays.
}

