% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_list.R
\name{restrict_list}
\alias{restrict_list}
\title{Restrict a list to the nodes satisfying a given property.}
\usage{
restrict_list(.list, .property = "is.array", .restrict = TRUE)
}
\arguments{
\item{.list}{The list-object to work upon.}

\item{.property}{A character, default \code{"is.character"} that
gives the name of the test function that tests the property of
the different nodes.  Note, the function must return a logical
value.}

\item{.restrict}{Logical value, default \code{TRUE}, that decides
if \code{.list} should be restricted to those nodes satisfying
\code{.property}, or if the restriction instead should be based
on the complementary nodes.}
}
\value{
A new list will be returned, with only those nodes of
    \code{.list} that satisfied the requirements given by
    \code{.property} and \code{.restrict}.
}
\description{
This function is mainly intended to simplify the task of working
with list where some of the nodes are arrays, but it could be that
it also could be of use in other cases.
}
