% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_sample.R
\name{seed_sample}
\alias{seed_sample}
\title{A helper when creating seed-values}
\usage{
seed_sample(n = 1, .seed = NULL, kind = NULL, normal.kind = NULL,
  vstr = NULL)
}
\arguments{
\item{n}{An integer that specifies how many seeds that should
be generated.  The default value is \code{1}.}

\item{.seed}{An initial seed to be used to ensure reproducibility
of the result from this function, and this value will be an
attribute of the resulting vector.  The default value for this
argument is \code{NULL}, and when left unchanged a value will
be sampled and used for this purpose.}

\item{kind}{Character or \code{NULL}.  If \code{kind} is a
character string, set R's RNG to the kind desired.}

\item{normal.kind}{Character string or \code{NULL}.  If it is a
character string, set the method of Normal generation.}

\item{vstr}{A character string containing a version number of
R. The default value \code{NULL} will use the present version
number.}
}
\value{
A vector of seeds between the highest and lowest available
    integer, i.e. between \code{-2147483647} and \code{2147483647}
    (the value used is the result of \eqn{2^31-1}.)  Attributes
    will be added in order to recreate the vector.  NB: If \code{n}
    is given as \code{1}, and \code{.seed} has been specified, then
    the returned value will be that value with the collection of
    attributes added to it.
}
\description{
This function creates a vector with one or more seeds, which often
will be needed in order to ensure that simulations done in parallel
are reproducible.  In addition some attributes will be added to
document the random number generator used to obtain the values.
See the help-page of \code{Random} for details about the three
arguments \code{kind}, \code{normal.kind} and \code{vstr}.
}
