% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_Namespaces.R
\name{the_Namespaces}
\alias{the_Namespaces}
\title{A function that returns information about the loaded namespaces.}
\usage{
the_Namespaces(.ns, .details = TRUE, .pattern, .type = c("all",
  "exported", "internal"))
}
\arguments{
\item{.ns}{An optional argument that can be used if it's of
interest to inspect a particular namespace.  The default is
that all the loaded namespaces are investigated.}

\item{.details}{A logical value that specifies the amount of
information to return, default value \code{TRUE}.}

\item{.pattern}{An optional regular expression. Only names matching
\code{.pattern} are returned.  \code{glob2rx} can be used to
convert wildcard patterns to regular expressions.}

\item{.type}{One of the values \code{c("all", "exported",
"internal")}, the default being \code{"all"}, which can be used
to restrict the returned details with regard to show only
exported or internal objects.}
}
\value{
If \code{.details} is \code{FALSE}, the result will simply
    be the names of the loaded namespaces (in essence the same
    result as obtained when using \code{sort} upon the result from
    \code{loadedNamespaces}).  If the \code{.ns}-argument has been
    supplied, then only the namespace specified there will be
    treated.  Different attributes will be added to the result when
    \code{.details} is given as \code{TRUE}.  One of the attributes
    will be \code{.pattern} (which mainly is included in order for
    the print-method to figure out what to write), and then there
    will in addition be one attribute corresponding to each
    namespace (having the same names), and each of these attributes
    will be a list containing the following components.

\describe{

\item{content}{A vector, that describes the content of the
    namespace; it's the result of \code{ls} used upon the
    namespace, with restrictions based on \code{.pattern}.  No
    restrictions will be used if \code{.pattern} is left
    unspecified.}

\item{exported}{A logical vector that for each part of
    \code{content} tells if the object was exported or not.}

\item{is}{A list that gives the result of \code{is} when it is used
    upon the different components.}

}
}
\description{
This function inspects the loaded namespaces, and returns
information in accordance with the user-defined specifications.
}
