% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_function.R
\name{this_function}
\alias{this_function}
\title{Capture information about the function above.}
\usage{
this_function(n = 1, details = c(0, 1, 2, 3, 4))
}
\arguments{
\item{n}{A positive integer that select which generations to
report.  The default value \code{1} selects the parent of
\code{this_function}, i.e. the function from which it was
called.  \code{2} gives the grandparent (needed when
\code{this_function} is called from within the
\code{error}-function), and so on.  An error will be returned
if \code{n} is out of range.}

\item{details}{A value that is compared against the values
\code{c(0, 1, 2, 3, 4)}, which decides the amount of additional
information to include as attributes.  The default will be to
use \code{0}.}
}
\value{
This function will return a result that depends on the
    values of \code{n} and \code{details}, where the first
    specifies which function to return whereas the latter one
    decides the amount of information included as attributes.  No
    attributes will be added when \code{details} is given as
    \code{0}.  Higher values of \code{details} adds attributes as
    described below.

\describe{

\item{info}{Attribute added when \code{details} is greater than or
    equal to \code{1}, which is a list that specifies the following
    details about the function that has been returned:
    \code{origin}, \code{package}, \code{status} and
    \code{arguments}.  The value in \code{origin} presents the
    place where the function was defined, which can be a namespace,
    the global environment or the name of another function.
    \code{package} is a logical value that is true when
    \code{origin} refers to a package.  \code{status} will be
    either \code{exported} or \code{internal} when the function
    originates from some package, it will be \code{global} if the
    function was defined in the global environment, it will be
    \code{local} if it was defined in some other function, or it
    will be \code{unknown} if the algorithm couldn't select one of
    the other ones (this might e.g. happen if a function has been
    defined in an environment that has been \code{attach}ed to the
    search-path.}


\item{lineage}{Attribute added when \code{details} is greater than
    or equal to \code{2}.  This is a list having names equal to the
    functions recorded from the call-stack, and each component of
    this list contains details as given for the attribute
    \code{info}.}

\item{sys.status}{Attribute added when \code{details} is greater
    than or equal to \code{3}.  A named list that gives the
    \code{sys.status}-information that was used to create the other
    attributes, see the help-page of \code{sys.parent} for the
    details.}

\item{content}{Attribute added when \code{details} is greater than
    or equal to \code{4}.  A named list that for each occurrence of
    a function adds a list with a component \code{objects}, which
    gives the content of the functions environment; and a component
    \code{is_test} that holds an array where these components has
    been tested against the \code{is.<something>} tests exported
    from the namespaces.  Note that the result listed here will be
    the status at the time \code{this_function} was called, and
    that the status of the environments might have changed if it's
    called at different times within the same target function.}

}
}
\description{
This function collects information about the function that it's
called from, based on the result from \code{sys.status}.  This
function can extract a variety of information, that other functions
can piggyback on instead of having to include large chunks of
similar code.
}

