% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_call.R
\name{update_call}
\alias{update_call}
\title{Update a call}
\usage{
update_call(.call, ..., .list = FALSE, .ignore = NULL)
}
\arguments{
\item{.call}{The call we want to update.}

\item{...}{\code{dotsMethods}, the arguments to be used for an
update of the formals of \code{.fun}.  If the arguments of interest
already are stored in a list (or an environment), set the argument
\code{.list} to \code{TRUE}.}

\item{.list}{Logic value, default \code{FALSE}.  This can be used
when the stuff that we want to update with already is contained in
a list or an environment. NOTE: Only the first argument given to
\code{...} will in this case be recorded!}

\item{.ignore}{Character vector, default value \code{NULL}.  When
the arguments we want to create our call are delivered in a list or
an environment, this can be used to perform a partially update,
i.e. give the names of the arguments we don't want to update (yet).}
}
\value{
The result of this function is that the formals of
\code{.fun} is updated at the calling environment.
}
\description{
The function \code{modify_call} in the \code{pryr}-package does not
always do as I expect when used inside of a function, so this
modified version of \code{update_formals} has been used in order to
get stuff to work in a way I better understand.  This approach
seems to be a bit saffer in those cases where it's desired that the
value \code{NULL} should be used as an argument.
}
