% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_abind.R
\name{my_abind}
\alias{my_abind}
\title{Bind arrays together, based on their dimension names.}
\usage{
my_abind(...)
}
\arguments{
\item{...}{A sequence of disjoint arrays (or \code{NULL}s, that
will be ignored) that we want to put together into a larger
array.  The arrays must have the same number of dimensions with
properly named dimension names.  The dimension names of these
arrays must pairwise have empty intersections.}
}
\value{
A larger array containing all of the smaller ones, put
    together by the help of \code{abind} along a dimension selected
    by the code.
}
\description{
This function is a wrapper for \code{abind} that binds a bunch of
arrays together after an inspection of their dimension-names (it
will crash if the the arrays have no named dimensions.  As this is
a wrapper for \code{abind}, it will be slightly slower, but the
effect is tiny for large arrays and it simplifies the writing of
the code in other functions.
}
